+++ js/util.js
@@ -1,0 +1,5 @@
+const checkStringLength = (string, length) => string <= length;
+
+checkStringLength();
+
+export { checkStringLength };
